portable convertFiles script :)

+++ scripts/netcdf-migration/convertFiles.sh
@@ -3,9 +3,9 @@
 processDirectory() {
   for file in *; do
     if [ -d $file ]; then
-      pushd $file
+      pushd $file > /dev/null
       processDirectory
-      popd
+      popd > /dev/null
     fi
   done
   processNCDFs
@@ -18,13 +18,20 @@ processNCDFs() {
       ad_accession=`ncdump -h $ncfile | grep ADaccession | psed 's/.\+"\(.\+\)".\+/\1/'`
       underscore=_
       new_name=$accession$underscore$ad_accession
-      echo "$ncfile => $new_name.nc"
-      ncdump -v BS2AS,DEacc,GN,ASacc,BSacc,EFSC,SC,SCV,EF,EFV,uVAL,uVALnum,PVAL,TSTAT,ORDER_ANY,ORDER_UP_DOWN,ORDER_UP,ORDER_DOWN,ORDER_NON_D_E,BDC $ncfile | sed "/double AS(AS)/d" | sed "/double BS(BS)/d" | sed "/double DE(DE)/d" > $new_name.data
-      if ncgen -o $new_name.nc $new_name.data; then
-        rm $new_name.data $ncfile
+      if [ "$ncfile" == "$new_name.nc" ]; then 
+        echo Skipping $ncfile
+      else
+        echo "$ncfile => $new_name.nc"
+        variables=`ncdump -h $ncfile | psed -e "1,/variables:/d" -e "/^$/,/}/d" -e "s/.\\+[a-z]\\+ //" -e "s/(.\\+//" -e "/^AS$/d" -e "/^BS$/d" -e "/^DE$/d"`
+        variables=`echo $variables | psed "s/ /,/g"`
+        ncdump -v $variables $ncfile | psed -e "/double AS(AS)/d" -e "/double BS(BS)/d" -e "/double DE(DE)/d" > $new_name.data
+        if ncgen -o $new_name.nc $new_name.data; then
+          rm $new_name.data $ncfile
+        fi
       fi
     fi
   done
 }
 
 processDirectory
+
